+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Add `getExpirationMonth` and `getExpirationYear` to `PaymentMethodNonceDetails`
+* Add `ExcessiveRetry` to `GatewayRejectionReason`
 
 ## 3.19.0
 * Deprecate `chargebackProtectionLevel` and add `protectionLevel` to `Dispute` and `DisputeSearchRequest`